from packets.c extracted packets_cheats.c and packets_input.c
